+++ bin/metadata_to_samplesheet.py
@@ -1,0 +1,1 @@
+../src/python/grz_qc/metadata_to_samplesheet.py
